requirements.txt
Django==4.2.13
matplotlib==3.6.3
pandas==2.0.3
plotly==5.22.0
Requests==2.32.3

django-environ~=0.11.2